added delegation engine init to vega node
